--- prowler/providers/gcp/lib/audit_info/audit_info.py
@@ -1,10 +1,0 @@
-from prowler.providers.gcp.lib.audit_info.models import GCP_Audit_Info
-
-gcp_audit_info = GCP_Audit_Info(
-    credentials=None,
-    default_project_id=None,
-    project_ids=[],
-    audit_resources=None,
-    audit_metadata=None,
-    audit_config=None,
-)
