--- prowler/providers/kubernetes/lib/audit_info/audit_info.py
@@ -1,9 +1,0 @@
-from prowler.providers.kubernetes.lib.audit_info.models import Kubernetes_Audit_Info
-
-kubernetes_audit_info = Kubernetes_Audit_Info(
-    api_client=None,
-    context=None,
-    audit_resources=None,
-    audit_metadata=None,
-    audit_config=None,
-)
