+++ tests/providers/kubernetes/kubernetes_provider_test.py
@@ -3,7 +3,15 @@ import sys
 from argparse import Namespace
 from unittest.mock import MagicMock, patch
 
-from prowler.providers.kubernetes.kubernetes_provider_new import KubernetesProvider
+from prowler.providers.kubernetes.kubernetes_provider import KubernetesProvider
+
+
+def mock_set_kubernetes_credentials(*_):
+    return ("apiclient", "context")
+
+
+def mock_get_context_user_roles(*_):
+    return []
 
 
 class TestKubernetesProvider:
@@ -157,3 +165,58 @@ class TestKubernetesProvider:
 
         roles = provider.search_and_save_roles([], [], "test-user", "ClusterRole")
         assert isinstance(roles, list)
+
+    @patch.object(
+        Kubernetes_Provider, "__set_credentials__", new=mock_set_kubernetes_credentials
+    )
+    @patch.object(
+        Kubernetes_Provider, "get_context_user_roles", new=mock_get_context_user_roles
+    )
+    def test_set_audit_info_kubernetes(self):
+        provider = "kubernetes"
+        arguments = {
+            "profile": None,
+            "role": None,
+            "session_duration": None,
+            "external_id": None,
+            "regions": None,
+            "organizations_role": None,
+            "subscriptions": None,
+            "context": "default",
+            "kubeconfig_file": "config",
+            "config_file": default_config_file_path,
+        }
+
+        audit_info = set_provider_audit_info(provider, arguments)
+        assert isinstance(audit_info, Kubernetes_Audit_Info)
+
+    def test_set_provider_output_options_kubernetes(self):
+        #  Set the cloud provider
+        provider = "kubernetes"
+        # Set the arguments passed
+        arguments = Namespace()
+        arguments.quiet = True
+        arguments.output_modes = ["csv"]
+        arguments.output_directory = "output_test_directory"
+        arguments.verbose = True
+        arguments.output_filename = "output_test_filename"
+        arguments.only_logs = False
+        arguments.unix_timestamp = False
+
+        audit_info = self.set_mocked_kubernetes_audit_info()
+        mutelist_file = ""
+        bulk_checks_metadata = {}
+        output_options = set_provider_output_options(
+            provider, arguments, audit_info, mutelist_file, bulk_checks_metadata
+        )
+        assert isinstance(output_options, Kubernetes_Output_Options)
+        assert output_options.is_quiet
+        assert output_options.output_modes == ["csv"]
+        assert output_options.output_directory == arguments.output_directory
+        assert output_options.mutelist_file == ""
+        assert output_options.bulk_checks_metadata == {}
+        assert output_options.verbose
+        assert output_options.output_filename == arguments.output_filename
+
+        # Delete testing directory
+        rmdir(arguments.output_directory)
